Moved airgap detection test from integration tests to airgap module
