+++ backend/tests/utils/mod.rs
@@ -5,14 +5,14 @@ use std::net::SocketAddr;
 use sqlx::SqlitePool;
 use tokio::net::TcpListener;
 
-use abacus::{airgap::AirgapDetection, router, shutdown_signal};
+use abacus::{router, shutdown_signal};
 
-async fn serve_api_inner(pool: SqlitePool, airgap_detection: AirgapDetection) -> SocketAddr {
+pub async fn serve_api(pool: SqlitePool) -> SocketAddr {
     let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
     let addr = listener.local_addr().unwrap();
 
     tokio::spawn(async move {
-        let app = router::create_router(pool, airgap_detection).unwrap();
+        let app = router::create_router_without_airgap_detection(pool).unwrap();
 
         axum::serve(
             listener,
@@ -24,32 +24,4 @@ async fn serve_api_inner(pool: SqlitePool, airgap_detection: AirgapDetection) ->
     });
 
     addr
-}
-
-pub async fn serve_api(pool: SqlitePool) -> SocketAddr {
-    let airgap_detection = AirgapDetection::nop();
-
-    serve_api_inner(pool, airgap_detection).await
-}
-
-pub async fn serve_api_with_airgap_detection(pool: SqlitePool) -> SocketAddr {
-    let airgap_detection = AirgapDetection::start(pool.clone()).await;
-    let mut violation_detected = false;
-
-    for i in 0..=200 {
-        if i == 200 {
-            panic!("Airgap detection failed to detect violation after 10 seconds");
-        }
-
-        if airgap_detection.violation_detected() {
-            violation_detected = true;
-            break;
-        }
-
-        tokio::time::sleep(tokio::time::Duration::from_millis(50)).await;
-    }
-
-    assert!(violation_detected, "Airgap detection did not run");
-
-    serve_api_inner(pool, airgap_detection).await
 }
